+++ CHANGELOG.md
@@ -16,3 +16,12 @@
 * fixed :sinkmode op
 * defaggregateop no longer fails when returning nil
 * Better error reporting on parametrized ops (variadic args aren't supported; this now throws an error.)
+* MUCH better error messaging on failed tests. Here's the old way:
+
+    expected: (= (map multi-set (map doublify set1)) (map multi-set (map doublify set2)))
+      actual: (not (= ({[2.0] 1}) ({[1.0] 1})))
+
+vs the new:
+
+    expected: (= input output)
+    actual: (not (= [[2.0]] [[1.0]]))